Add initial support for Swift registry (#1180)
This is the initial implementation of the Tuist Package Registry.

There are a couple of pieces missing, primarily:
- `Package.swift` specific to language versions (e.g.,
`Package.swift@6`)
- Authentication
- Cost attribution

As for the implementation, there are three important pieces:
- `UpdatePackagesWorker` – worker that polls the list of packages from
SwiftPackageIndex (more specifically, from
[here](https://github.com/SwiftPackageIndex/PackageList/blob/main/packages.json))
and adds those (and their releases) that are missing
- `UpdatePackageReleasesWorker` – worker that goes through our database
and updates releases of packages that were the last updated using the
`last_updated_releases_at` column.
- `swift_controller` – controller that implements registry endpoints
specified
[here](https://github.com/swiftlang/swift-package-manager/blob/main/Documentation/PackageRegistry/Registry.md#4-endpoints)

Each worker has a dedicated token – however, otherwise, we don't exactly
how many calls will each worker make. This is something we'll need to
tinker with or implementing a moving window that would track API calls
made by a given token in the last hour. However, it's quite likely that
after the initial warmup, it won't be that likely that we'll hit the API
limits even if we update packages relatively often (at least once per
day).

You can check out the tables getting populated on the deployed staging:
-
[packages](https://supabase.com/dashboard/project/inzgspjesrhqhleomvkb/editor/107988?schema=public)
-
[package_releases](https://supabase.com/dashboard/project/inzgspjesrhqhleomvkb/editor/108018?schema=public)

There are some issues still popping up here and there, so I will be
updating the PR as I discover them. However, there are no larger changes
planned, so this is ready for review.