Introduce Boundary (#1205)
In Elixir, any module can depend on the public interface of another
module. Without control this can lead to an architecture that's hard to
reason about and maintain, and slow compilations due to circular
dependencies. Therefore, I propose that we introduce a system to declare
and enforce an architecture in our codebase.

This PR introduces [boundary](https://hexdocs.pm/boundary/Boundary.html)
for that, and defines two boundaries, `Tuist`, and `TuistWeb`. In a
follow-up PR I'll define nested boundaries within `Tuist`. For example,
accounts can't on the swift package registry context.