Do not recreate packages with vX.Y.Z prefix (#1215)
Resolves: https://github.com/tuist/server/issues/1214

There's an issue when updating package releases where we incorrectly try
to recreate an existing package release when the release is in the
format of `vX.Y.Z` instead of the plain `X.Y.Z`.

The issue occurs because when filtering releases we don't have in the
database, yet, we compare the DB version with the tag – but the DB
version is already the semantic version (without the leading `v`).

Fix is to compare only semantic variants of version tags.