Fix updating Package.swift when it references remote artifacts (#1218)
Resolves: https://github.com/tuist/server/issues/1219

We are currently doing a regex to replace packages referenced by name
with `.product(...)`.

However, the regex is also taking in packages with remote artifact URLs
– when trying to convert those into scope and name, the server throws an
error.

We should only try convert URLs that are in the format of
`https://github.com/org/repo`. If a URL has more path components, we
should filter those out.