Create "TuistTest" boundary (#1225)
All the modules under the `test/support` folder belonged to the `Tuist`
and `TuistWeb` boundaries, causing boundary checks to throw warnings
because of invalid boundaries.

To mitigate those warnings, I'm creating a boundary with all those test
modules, `TuistTest`, with boundary checks disabled, and placing
utilities, cases, and fixtures modules under it.

```elixir
defmodule TuistTest do
  use Boundary, check: [in: false, out: false]
endf
```

Most of the changes in this PR are the updates of the references to
those modules to adhere to the new naming under the `TuistTest` module
boundary.

> [!NOTE]
> In the future, we might want to define boundaries from`TuistTest` to
`Tuist` and `TuistWeb`