Lower the limit of package releases to update per hour (#1228)
Resolves: https://github.com/tuist/server/issues/1227

I'm lowering the number of package releases that get updated per hour to
lower the chance of exhausting the API limit.

I don't want to lower the limit more than that as otherwise, we would no
longer update packages with new releases in <24h. If the error persists,
we might want to add an additional token to rotate.