(hotifx): Use dedicated accounts to update packages (#1229)
I've created two new GitHub accounts for updating packages and their
releases:
- `tuist-update-packages-worker`
- `tuist-update-package-releases-worker`

Turns out the 5k [hourly request
limit](https://docs.github.com/en/rest/using-the-rest-api/rate-limits-for-the-rest-api?apiVersion=2022-11-28)
is for the whole account. And when we exhaust the API limits of the
`tuist` account, our CI starts failing as well:
https://github.com/tuist/server/actions/runs/12479101004/job/34827790814

I also updated the workers to only periodically run in production, so we
don't exhaust our API limits because of non-prod environments.