Update outdated note about having to use SPM to run tuist
I've been running my local build through `tuist run` all day yesterday
and hit no obstacles, so there's no general blocker to do this.
If we, in the future, find that this does have some nuances around it
where in specific cases, SPM needs to be used, we can add this back for
the specific cases. For now, we are not aware of any.

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>