Add `PATCH /api/accounts/:handle` endpoint (#1236)
Server part of tuist/tuist#6666.

Adds a new endpoint to update a user's account, in this case only the
`name` (handle) being available to update.
This is pretty straightforward, at least until I realized that this
handle is stored locally as part of the authentication token. The
CLI-side implementation therefore includes updating the token. In its
default implementation, Guardian refreshes all Guardian-managed fields
(`iat`, `exp`, `jti`, ...) from the token, while copying over the other
claims unchanged. Unfortunately, it does this using a collection of
private helper functions, meaning that in order for us to be able to
change the `preferredUsername` claim, we have to reimplement those
helpers ourselves with the functions that are publicly available.

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>