Add ability to change account handle (#7209)
Adds a new account update command that allows updating a user's or organization's account.  
Currently, this supports updating the handle of the account using `tuist account update --handle {new_handle}`.

Since for user the handle is also stored in the authentication tokens that are persisted locally, the tokens are also refreshed upon update.