Elixir 101: Using patches in LiveView (#1245)
Both HTML anchor tags and `redirect` result in full page reloads. In
LiveView terms, that means
1. Closing the existing WebSocket connection and opening a new one.
2. Rerunning the plug pipeline (which, due to authentication, causes
database hits).
3. Rerunning the `mount/3` callback.
4. Doing a full page re-render.

In user terms, it means losing all input (which we don't have at the
moment, so no impact), and a full render flash.
The `.link` component with the `patch` attribute or the `push_patch/2`
function should be used when changing parameters _in the same LiveView_,
like in this case. We're staying on the same page, with the same
authentication, and are only changing filter parameters. Patches in
LiveView result only in the `handle_params/3` callback being called, and
a partial re-render happening, with only the diff being sent to the
client. This is both less resource-intensive and a nicer experience for
the user.

Before and after visual:


https://github.com/user-attachments/assets/9dd633bc-06f8-4ca7-b7b9-85ab6ebab598

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>