Allow users to create account tokens (#1267)
Currently, only admins can create account tokens.

However, one needs to be able to create an account token to log in to
the registry (`tuist registry login`). Since account tokens can't really
be used for anything else but reading the registry, it's fine for users
to be able to create one.

Long-term, as we add more capabilities to the account token (such as
when we merge it with the project token), we'll need to make sure that
the user creating that token doesn't create an account token that has
more permissions than the user has (we'll want to follow how GitHub's
fine-grained tokens work).