Add an index to command_events.preview_id (#1262)
This is our slowest database query by multiple orders of magnitude, due
to the size of the table and the lack of an index on `preview_id`.

The `comment_events` table is a TimescaleDB Hypertable, which does not
support foreign keys, resulting in us using a raw value for the `where`
query. Since it cannot use the foreign key index, we're adding one
manually in this change.

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>