Increase the security of the server by using the Content-Security-Policy header (#1263)
Resolves https://github.com/tuist/server/issues/1241

We are not returning the `Content-Security-Policy` from the server, and
thus not leveraging one of browsers' security capabilities to control
which resources can be loaded.

This PR adds sets that up, and configures the header to work with our
existing needs. I reused the same policy across the marketing and the
app site to keep things simple initially, but down the line we might
want to split that into two different policies.

> [!NOTE]
> Going forward make sure that if you add third-party resources, or
inline scripts or styles, you might need to tweak the configuration or
use the `nonce` attribute.