Do not replace products in nested Swift package manifests (#1281)
Some packages, like https://github.com/danger/swift, have a nested
`Package.swift` file which is _not_ the Swift Package manifest that's
used to define a Swift package. Such file can only be in the root.

However, as of now we'd try to update a nested `Package.swift` to fix
some issues in SwiftPM.

This PR fixes that, so we correctly update only Swift package manifests
in the root of the project.