Double the prod vm memory to fix OOM errors when updating releases (#1280)
I'm doubling the prod VM's. We're going through the unzipped archives in
memory and some packages, like https://github.com/rhx/swiftgtk, turn out
to have quite large directories (around 8 GB which was our current
limit).

Long-term, we might want to consider moving to FLAME when processing
packages to decrease the impact we might be having on the rest of the
application.