Add support for pre-release package releases (#1274)
As per [semver](https://semver.org/), versions can also have pre-release
and build identifiers such as:
```
5.0.0-beta
5.0.0-beta.6
```

To define a dependency on a version such as `5.0.0-beta.6`, you'd add
the following entry in your `Package.swift`:
```swift
      .package(
            id: "groue.GRDB_swift",
            exact: .init(
                5, 0, 0,
                prereleaseIdentifiers: ["beta"],
                buildMetadataIdentifiers: ["6"]
            )
        )
```

There is a slight discrepancy between SwiftPM's definition of semantic
versions where the delimiter between metadata identifier and a
prerelease is a `+` instead of `.`. In this case, it would try to
resolve the version of `5.0.0.-beta+6`. When saving a version in the
database, we thus need to also do this replacement.