Implement the newsletter issue template (#1233)
[Handbook PR](https://github.com/tuist/handbook/pull/50)

I'm adding the business logic to render newsletter issues and the atom
and RSS feeds to consume the issues via RSS/Atom readers. I ended up
defining a custom schema for issues in a `.yml` file because there was a
considerable markup duplication across issues and I don't expect the
schema to change frequently. If it did, we can introduce those changes
incrementally.

---------

Co-authored-by: Marek Fořt <marekfort@me.com>