Noora: Add Dropdown component (#1305)
Adds the Dropdown component from the Noora design system.


[Figma](https://www.figma.com/design/AOrO3lGZvGvtD6hA6nIQmb/Noora-Design-System?node-id=683-2447&m=dev)

## Implementation

This is our first interactive component that goes deeper than "banner
close button". With interactive components, there's *a lot* to get
right:
- Focus trapping
- Keyboard navigation (and highlighting that works for both keyboard
selection and hover)
- Typeahead (selecting items by typing their label)
- Accessibility labels (for dropdown, that's open/closed, labels,
activedescendant for highlights, and a lot more)
- Positioning (Where does the dropdown open, up top, below, left, right,
what happens if it's on the edge of the page...)

Writing that ourselves is... honestly unrealistic. We lack both the
expertise and time to try to fully implement that. This is why I am
introducing [ZagJS](https://zagjs.com) in this PR. Zag is a headless UI
component library that for a change is not bound to a frontend
framework, and works with regular JavaScript, and, through some hook
magic that I wrote, Phoenix and LiveView. I've spent a lot of time
researching component libraries to make work with Phoenix, together with
a friend who is a frontend engineer, and Zag is *really* good. Then I
spent a bunch more time writing the glue glayer between LiveView and
Zag.
For visualisation, [this](https://state-machine-viz.vercel.app/menu) is
the state machine we're getting for free for the dropdown.

This is a bunch of extra JavaScript we're loading, but it's also _a lot_
of functionality. In addition, a bunch of the loaded code is
`@zag-js/core`, which is shared across all components and which we will
reuse for the modal, tabs and other components as we get further.

## Screenshots

![CleanShot 2025-01-20 at 13 22
57@2x](https://github.com/user-attachments/assets/f9cdd05f-be69-4462-a12b-58602106404b)
![CleanShot 2025-01-20 at 13 23
14@2x](https://github.com/user-attachments/assets/bf2b0171-96b3-4190-b708-7441c746301a)
![CleanShot 2025-01-20 at 13 23
40@2x](https://github.com/user-attachments/assets/8e0775f6-981d-496b-ba7b-3012aea52d3a)

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>
Co-authored-by: Marek Fořt <marekfort@me.com>