Surface module hashes in dashboard by storing Xcode Graph for each run (#1334)
Related: https://github.com/tuist/tuist/issues/7269
Counterpart CLI PR (currently in draft):
https://github.com/tuist/tuist/pull/7278

This PR surfaces the module hash in the dashboard:

![image](https://github.com/user-attachments/assets/bda66ca8-d2c4-42be-9ae6-4bc62b2d00e9)

This required a couple of changes. Until now, the target-level data for
binary cache and selective testing was reported by explicit parameters
in the API and their counterpart columns in the `command_event`:
- `cacheable_targets`
- `local_cache_target_hits`
- `remote_cache_target_hits`
- `test_targets`
- `local_test_target_hits`
- `remote_test_target_hits`

All of these API parameters are now deprecated and the columns should be
eventually removed once most of our customers move to the version of the
CLI that has implemented the new approach.

In the new approach, instead of attaching all those specific fields at
the `command_event` level, we store parts of the `xcode_graph` in the
database:
- `command_event` owns a new `Graph` model
- New `Graph` model owns a new `Project` model
- A new `Target` model is owned by `Project`. The `xcode_graph_targets`
table then has the extra binary cache and selective testing-related
metadata:
  - `binary_cache_hit`
  - `binary_cache_hash` (completely new data)
  - `selective_testing_hit`
  - `selective_testing_hash`

We have considered also storing the full graph in a JSON column stored
in the `command_events` table. However, because we expect to query the
hashing data in a not-so-distant future, we went with storing the
relevant parts of the graph in the table.