Use XcodeGraph when inserting test cases (#1338)
This follows up the work done to upload the Xcode graph to the server,
and uses it for analytics instead of calculating module hashes on the
client and sending them to the server.

Additionally, it reactivates the test case insertion, behind a feature
flag. This PR does not include any performance improvements (and likely
makes things slower for the time being), so keep this in mind if ever
activating the feature for end users until we land some improvements
there.

Co-authored by @fortmarek

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>