Add .icon component (#1369)
Previously, we only generated components as functions with the icon name
equalling the function name. This should be the preferred usage when
building pages and passing icons inside slots, but is limiting when we
need to pass an icon _inside_ a slot. In particular, LiveView allows

```
slot :head do
  attr :icon, :string
end
```

but does _not_ allow

```
slot :head do
  slot :icon
end
```

To allow us to use icons as strings, this PR adds a helper component
that can be used as `<.icon name="user" />`.

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>