Noora: Input Field (#1358)
Welcome to another rendition of "rate the API I came up with".

<img width="464" alt="image"
src="https://github.com/user-attachments/assets/1dd9207b-602b-4c94-92f3-485cf62d97a6"
/>



[Figma](https://www.figma.com/design/AOrO3lGZvGvtD6hA6nIQmb/Noora-Design-System?node-id=967-2148&m=dev)

Some implementation notes:
- I consciously left out the "Invite members" type for now. Adding a
dropdown inside the input would be quite a bit of work and figuring out
how to handle it inside a form, so I'd like to wait for us to fully
figure out permissions on user level before going there.
- For the hint suffix, I've made it so that it automatically integrates
a tooltip. It's the cleanest surface area in my opinion, and it can
still be overridden by a completely custom suffix by passing a slot.
- The credit card icon seems to be completely custom? I've used the one
from Tabler for now, but am happy to update it. If it's custom and
supposed to be, should we toss it into the same svg folder?
- I've went with the `tel` type for card numbers, as that's what
everyone on the internet says is best.
- The suffix for `search` assumes that cmd+k is always the search
shortcut. Let's go back there when we actually implement keyboard
shortcuts :)

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>