Noora: Refactor existing components (#1385)
I don't recommend reading through the code line by line, as it's just
shuffling things around. Having the before and after storybook side by
side helped me to make sure everything is the same, but I'm sure I
missed a thing or two.

- Adds documentation to each module
- Uses border tokens where applicable
- Uses data attributes for part identification instead of classes to
match implementation of interactions
- More CSS variables and fewer direct assignments
- Structures and sorts elements using nesting

None of this should have any visual changes or impact on behavior
(except for when the font shorthand was broken, that's fixed now)

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>