Add a "community" page and blog post introducing our first ambassadors (#1400)
Having a page that recognises our most active contributors is something
that I've been meaning to do for some time so with more contributors
joining, specially in the translation front, I thought it was the
perfect time to put that page together. This first iteration of the page
is quite simple, just a markdown page with sections and lists in each of
them. In the future we can iterate on it to include avatars of the
people.

Additionally, I included a blog post where I announce our first
ambassadors.

---------

Co-authored-by: Marek Fořt <marekfort@me.com>