Test only TuistExtKitCanaryAcceptanceTestWithIOSAppWithFrameworks test suite (#1404)
Running _all_ of our canary acceptance tests is quite wasteful,
especially as the test suite keeps growing. It now takes over half an
hour to run these even when using the binary cache – and the deployment
to prod only happens after the tests are finished.

I'm scoping running the tests only to
`TuistExtKitCanaryAcceptanceTestWithIOSAppWithFrameworks` that tests
remote binary cache and selective testing work. That's the most critical
piece. The rest of the tests mostly retest the same endpoints and are
not critical, so we can skip them here.