Raise the pool size to align with the max client connections in Supabase (#1415)
I noticed that we were still getting timeouts despite our operations
being fast. After some digging, and with the help of [this blog
post](https://jpcamara.com/2023/04/12/pgbouncer-is-useful.html) to
better understand how connections work with Postgres through PgBouncer
(which we are using in our transaction connection mode), I noticed our
pool size was too low.

According to Supabase, with our database specs, we have the following
max client connections:
- Tuist: 800 connections
- Tuist Staging: 200 connections
- Tuist canary: 400 connections

If got it right, "client connections" refer to the connections between
the client (Ecto in our server) and PgBouncer. Therefore, the size of
the Ecto pool should be close but not above the "max client connections"
in Supabase.

If the assumption is correct, I believe this should mitigate the
timeouts we were seeing.

> [!NOTE]
> I left an extra buffer of 10, one of which is used during migrations.

> [!IMPORTANT]
> This configuration is valid with one machine per app. I checked in
Fly, and that's currently the case. If we wanted to introduce more
machines, for example for redundancy reasons, we'd need to tweak those
values to distribute the client connections across machines.