Adjust the pool size (#1416)
After merging [](), I noticed we got some `Max client connections
reached`
[errors](https://appsignal.com/tuist-cloud/sites/66967a1394809a7074e5bc90/exceptions/incidents/177)
after deploying to canary.

It turns out that I did not account for the fact that at deployment
time, there's a period during which there are 2 instances of the app
running. Therefore, we need to divide the number by two.

> [!NOTE]
> A better solution would consist of dynamically adjusting the size once
the deployment is done. I'll have to investigate if we can detect that
and use the information to adjust the Ecto pool size at runtime.