Remove the migration pool size limit (#1417)
Deployments
[fail](https://github.com/tuist/server/actions/runs/13497722822/job/37708670886)
because we limit the pool in migration to 1, and that causes timeouts in
the Ecto pool. I don't see why we need to set the limit to 1, so I'm
removing that limit.

The usage of connections during a deployment should look like this:

![Untitled Diagram-Page-3
drawio](https://github.com/user-attachments/assets/af8a114a-a9e8-45d3-adb5-7cda55b56d07)