Fix Hackney version to mitigate AppSignal issue (#1428)
It turns out that [the
issue](https://github.com/appsignal/appsignal-elixir/issues/981) is
caused by a bug in the underlying Hackney package so the downgrade of
AppSignal doesn't fix the issue. The
[logs](https://fly-metrics.net/d/fly-logs/fly-logs?orgId=99814&var-app=tuist-cloud&var-query=appsignal&var-instance=All)
in production reveal that the NIF doesn't load:

<img width="1132" alt="image"
src="https://github.com/user-attachments/assets/5ce47ef6-9cc1-439e-9bf9-a7c143ac99d2"
/>


I tried to update the Hackney version to [this
one](https://github.com/benoitc/hackney/releases/tag/1.23.0), which
fixes the issue, but the requirements of appsignal-elixir doesn't allow
it, so until `appsignal-elixir` releases a new version, I'm fixing
Hackney to the version we know doesn't have the bug.