Update database configuration for staging (#1430)
This continues @pepicrft's exploration.
Also, this was already manually deployed to test things out.

1. This reverts #1417. Given that migration machines are spawned as an
additional machine, having the migration machine have the same database
pool size essentially means we're losing 1/(number of app machines +1)th
of database connection limits that we have to keep open for the next
deployment.
2. It sets the pool size to 20.

## Calculating the right pool size

`tuist-staging` runs on the Micro Supabase compute module, which comes
with a default `max_connections` of 60.
Out of the 60,
- 1 is reserved for Supabases RLS authentication watcher
- 1 is reserved for backups
- 1 is reserved for migrations (see above)
- a handful should be reserved for connecting to the database manually
to unbreak things if needed

For this particular case, I decided to allow fourty connections for app
machines. Since we're running two concurrently, this means twenty
connections per machine.

## Resource usage

Moving from 5 to 20 connections has increased our memory usage in
Supabase from 40% to 62%. This is still well within "we're running out
of memory" levels, so I'm happy to set this and will continue to monitor
whether there are any more spikes.

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>