Add ran_at column for better accuracy of runs analytics (#1441)
Related: https://github.com/tuist/tuist/issues/7348

I'm adding a new `ran_at` column and adding it as a new param when
creating a command event (aka run). Since runs are not sent immediately
when they finish in local environments, the `created_at` column, which
has a value of when the item is inserted in the db, can represent date
that's minutes, hours or days after the command was actually run.

I did consider leveraging the `created_at` column, but storing when the
item was actually inserted at felt like something worth storing, too –
so an explicit column is imho a better fit.

See the client PR for more context (currently in draft):
https://github.com/tuist/tuist/pull/7372