Bundle frontend assets using `esbuild` (#1440)
## Context
Ad blockers blocked a CDN-served module that contained "cookie" in the
name. This prompted us to revisit our approach to how our markup
consumes CSS and JS.

## Approach
Leveraging the platform (browsers) to load CSS and JS is great, but it
falls short in a few areas:
- The above issue
- We can't leverage dependency-updaters
- Offline coding is not possible

## What
I've brought back `esbuild`, and use it to create 4 bundles for
marketing, app, api docs, and storybook routes. Transitive dependencies
of those bundles, like Noora, which could be bundled separately, is
duplicated, but we can optimise this down the line using `esbuild` (it's
supported).

I've then removed all the CDN-delivered assets in favour of
bundled-from-node-modules copies, and introduced a `package.json` along
with `pnpm` to resolve the dependencies at `mise install` time. I
reconciled those changes with our CI pipelines and the `Dockerfile`.

## How to test

Try to load marketing, storybook, api docs, and app routes in your
browser, and make sure they load successfully with no error in the
developer console.