Include the missing price_1QPK3YLWue9IBlPS3gmVGJin price item in the enterprise's flat_monthly items (#1498)
One of our soon-customer (account id 1583), which we assigned to an
enterprise plan on Stripe, did not get the subscription set in the
database. It turns out that we messed up a bit with price items in
Stripe for the enterprise product, and ended up with two price items,
one for Dollars, and another for Euros. Since our list of price items
missed one of them, the one associated to that customer, the
subscription didn't get set.

This PR adds the missing item. After we deploy this, I'll re-sync the
Stripe data for that customer to get the subscription in our end.