Increase DB queue target (#1502)
Due to the design of the cache API, which leads to sudden spikes in
requests hitting our server, our DB connection pool queue target and
interval teams lead to Ecto dropping database connections.

To prevent those drops, I'm increasing the queue target time to a bit
more than double of the current value. In a service with more stable
traffic, you wouldn't do this because you can make some slow queries go
unnoticed, but with the tools that we have in place (e.g. Supabase,
AppSignal), this should not hopefully not happen to us.

As a follow-up, I'm going to see if we can have access to the number of
connection dropped to see if there's a correlation between that and the
number of requests that our server is getting.