feat(noora/chart): Set colors and support light/dark mode (#1486)
This PR allows referencing CSS variables in chart options, as well as
supporting light and dark mode both with basic media queries and
`light-dark()`.

As such, this CSS works:
```
:root {
  --chart-1: "#FFF";
  --chart-2: light-dark("#F00", "#00F");

  @media (prefers-color-scheme: dark) {
     --chart-1: "#000";
  }
}
```

This can then be referenced in the chart options as `var:chart-1` and
`var:chart-2`.
These references work
- in the top-level `colors` configuration which creates the ECharts
theme
- any top-level axis, item and area style properties
- any axis, item and area style properties nested in a series

Along with the changes to enable this, I've updated the default
configuration as well as the stories. I used the test run story as a
nice example for the "CSS variables work deeply nested in series data
items" situation.

Lastly, I've reorganised the chart file so now it... makes sense.

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>