Improve our cookie consent (#1507)
The current cookie consent is a binary decision for all the users,
regardless of the legal framework of the country they are visiting from.
Because of that, Koala, which we use to derive the visitors' employer,
has much less data.

This PR changes our approach to use the consent provided by getterms.io,
whose behavior varies based in the jurisdiction of the visitor's
country. For example, in Europe, where we have GDPR, the user will see
the banner, but in other locations where there isn't any framework, they
won't see the banner.

Last but not least, this banner doesn't allow clicking for rejection.
You can either accept them, or configure them, and through the
configuration menu you can opt out.