Rename fly.toml to fly.prod.toml (#1511)
With everyone having access to our production instance, which I believe
it's the right thing to do at this stage, it's easy to mess things up
because `flyctl` commands default to the production instance due to the
name of the configuration file, `fly.toml`.

This PR renames the file to `fly.prod.toml` to eliminate that default
and force the caller to think about the environment that they want to
interact with.