Fix conflicting duplicated `secret_key` keys in the encrypted credentials (#1525)
When I moved `secret_key.base` from Fly secrets to the encrypted
credentials file, I didn't notice there was already a `secret_key`
attribute at the root of the credentials with a different value for
`secret_key.base`. This PR removes the new key that I added and adjusts
the value of the existing one to the one that we had in Fly secrets.

> [!NOTE]
> The failing deploy due to acceptance tests failing was caused by
trying to use a secret key base other than the one that was used to
encrypt the passwords.