Configure Hammer to be distributed (#1527)
While updating [Hammer to
v7](https://github.com/tuist/server/pull/1518), which we use for
rate-limiting, I noticed our rate-limiting is scoped to the host. This
is fine at the moment, since we only have one host, but towards
introducing redundancy in the future, it's better if we set up the
project to work in a distributed manner.

Hammer has some
[guidelines](https://hexdocs.pm/hammer/distributed-ets.html) to use
Phoenix PubSub to enable a distributed setup, so I followed them in our
setup.