feat(noora/card): Add card component (#1519)
Resolves: https://github.com/tuist/server/issues/1487

Adds the card component required for most, if not all, of our prototype
pages.

![CleanShot 2025-03-17 at 09 45
54@2x](https://github.com/user-attachments/assets/5d11471a-bb82-433e-b052-bdc17f8aed05)

The wrapper component is pretty straightforward since it's consistent
between uses. The section component I kept super generic, since in the
existing prototypes alone we have
- tables
- graphs
- multiple sections side-by-side as sort of overview

So this is gonna be pretty use-case and page specific. That's why I kept
the styling minimal (background, border, and a default padding).

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>
Co-authored-by: Marek Fořt <marekfort@me.com>