feat(noora/sidebar): Add sidebar (#1513)
Adds the initial sidebar components and setup for the dashboard
redesign.

- Moves all the button styles from outside the `--noora-button` scope so
we can reuse them.
- Adds a new `NooraCollapsible` hook.
- Adds the vertical version of the Tab menu component.
- Adds reusable, configurable sidebar components.
- Adds a new sidebar that is rendered when the `noora` feature flag is
active.

This does not yet add the mobile sidebar as I'm trying to figure out a
thing or two there, but I'm almost there. Already adds a non-collapsible
group configuration that I'll use for that.
The active item detection doesn't work yet as... the pages don't exist
yet and cannot be active. We will add these once we build out the pages
using the logic built out in
https://fly.io/phoenix-files/liveview-active-nav/.

![CleanShot 2025-03-14 at 17 37
30@2x](https://github.com/user-attachments/assets/82792915-40f4-43c2-b789-3ee957c11c24)

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>
Co-authored-by: Marek Fořt <marekfort@me.com>