Cache the authorization in `/api/cache` endpoints (#1529)
Bursts of requests going to `/api/cache` can exhaust our DB connection
pool leading to connection drops. Those bursts also caused issues in the
past, in particular CPU consumption spikes because we used `bcrypt` in
every request to authenticate the user, and that is a CPU-bound
operation.

Back then we solved it by caching authentication. The first request that
we got on the server triggered the authentication, and the rest waited
for this one to complete to get the cached result. I propose we do the
same thing for authorization with a time window of 1 minute, which
should be sufficient to alleviate some of the DB load when projects hit
our API.

> [!NOTE]
> Requests waiting should translate in a bit of extra memory
consumption, but I believe it should be insignificant.

> [!NOTE]
> Those endpoints have still some DB operations that we can't cache
(i.e. writing events). As suggested by @cschmatzler, I suggest that we
use [Broadway](https://elixir-broadway.org/) to ingest the data in
batches. This would also allow us to reduce the number of DB operations,
from one per request, to 1 query per batch including multiple command
event insertions.