fix: bad request if analytics endpoints are used authenticated as a project from a non-CI environment (#1539)
Resolves:
https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/665

The current implementation design makes a strong assumption that project
authentication means CI, and user-authenticate means non-CI, and we have
validations and designs based on that assumption. When the assumption is
not met, for example when a project tries to use a project token outside
of a CI environment, the validation fails causing a 5xx error from our
API.

I'm handling that gracefully by returning an error that the CLI will
show to the user. Down the road, we'll break that assumption by moving
to "account tokens" and decoupling it from the environment from where a
certain command ran.