feat(noora/checkbox): Add checkbox (#1535)
https://www.figma.com/design/AOrO3lGZvGvtD6hA6nIQmb/Noora-Design-System?node-id=540-2748&m=dev

Adds the checkbox component.
![CleanShot 2025-03-20 at 09 17
03@2x](https://github.com/user-attachments/assets/4e92b5f1-f30b-4073-9ef5-e9d21438c6ff)
![CleanShot 2025-03-20 at 09 17
38@2x](https://github.com/user-attachments/assets/93f086fc-7123-4d7d-9ee4-3dae0a23be8a)


Doesn't include the link slot for now or option to flip them as we don't
have it in our design at the moment, and this is the minimum
implementation that gets us forward.
Also, the indeterminate toggle is broken in Zag - we also don't use this
so it's not a big deal, but I had already added the styles before
noticing so I'm leaving it for the time being. Will be fixed when we
upgrade to Zag 1.x.

---------

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>