Drop requirement to upload command_event to display previews (#1540)
As mentioned
[here](https://github.com/tuist/tuist/pull/7418#discussion_r2005080322),
we want to drop the requirement of `command_event` to be uploaded in
order to show previews in the list of previews. To do that, we need to
add git-related metadata such as git branch directly in the `previews`
table.

Not having to list over command events but instead dealing with previews
directly made a lot of logic simpler as well, so that's a win.

CLI counterpart PR: https://github.com/tuist/tuist/pull/7420