Use single key when caching `/api/cache` data (#1567)
I misinterpreted how keys work with the `Cachex.transaction!` API. Each
item in the list represents a key under which a value will be stored,
and not the components that will form a unified key. Therefore, we were
storing values under multiple keys, which is not necessary. This wasn't
an issue until I merged the PR that configures [Cachex to be
distributed](https://github.com/tuist/server/pull/1560), which surfaced
the problem because distributed setups only support a single key:

> In the case you're using an action which is based on multiple keys
(such as
[Cachex.put_many/3](https://hexdocs.pm/cachex/Cachex.html#put_many/3) or
[Cachex.transaction/3](https://hexdocs.pm/cachex/Cachex.html#transaction/3)),
all keys within **a single call must live on the same destination
node**. This should not be surprising, and is similar to the likes of
[Redis](https://redis.io/) where this is also the case (at the time of
writing). If you attempt to use these types of calls with keys which
slot to different nodes, you will receive a :cross_slot error. As it's
typically difficult to guarantee that your keys will slot to the same
node, it's generally recommended to only call these functions with a
single key when used in a distributed cache (and so put_many/3 is then
redundant).

> [!NOTE]
> This wasn't caught by unit tests because we don't have a distributed
setup of nodes there. It was caught by acceptance tests though, showing
that they play a critical role in our deployment pipeline.

---------

Co-authored-by: Marek Fořt <marekfort@me.com>