Revert Cachex revert, fix its usage, and further optimize the `/api/cache` requests (#1571)
The reason why we experienced runtime issues on production is because
there were still some direct usages of `Cachex.transaction!` with
multiple keys. I've gone through all of them ensuring we use
`Tuist.Cache.get_value`.

Moreover, while doing some testing, I noticed there were still some
queries that happened because we have some `preload` calls to load the
account associated to the resource. I moved those to the loading of the
authenticated subject so that we don't need a second query to fetch the
associated account, since this is something we'll need in most cases.

> [!NOTE]
> There's one final optimization that we can do using Broadway to batch
the insert of all the download events, which we currently do 1 per
request to download or upload artifacts to the cache. I'll do that in a
follow-up PR.